make blumpkin coverage only average files that have lines in them
before blumpkin coverage was averaging the average for each file
this means that if you had an __init__.py file with 0 lines in it
it would have a coverage of 0, and then this would get averaged into
the overall project.  With this change, it makes the balanced resources
coverage go from 90% -> 98% since it is no longer averaging all the
files with zero for coverage.